requirements.txt
PyPDF2>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pdf2image>=1.16.0
pytesseract>=0.3.10